refactor(312 PR-A): relocate REPL UI (repl + renderer) → reyn.interfaces.repl
#312 stage PR-A — move the console/REPL UI modules #1700 left behind in
reyn.chat into the interfaces/ tree. Clean break (no shim), byte-gate.

- git mv: chat/repl.py + chat/renderer.py → reyn/interfaces/repl/ (100%
  rename, byte-identical). New reyn.interfaces.repl package.
- Repoint importers: cli/logger_factory (renderer), cli/commands/chat (repl),
  + a stale path-ref comment in intervention_choices. repl's internal
  `.registry` → absolute `reyn.chat.registry` (registry stays in the runtime
  pkg; interface→runtime is the correct direction). `.renderer` stays relative
  (renderer moved too). renderer's `reyn.chat.outbox` import unchanged (outbox
  stays in runtime).

SCOPE CORRECTION vs the FP-0044 cluster table: error_format.py is NOT moved —
it is runtime error-CLASSIFICATION ("Friendly classification of router/LLM-call
failure exceptions"; imports reyn.runtime.budget.BudgetExceeded; imported by
session.py's router-loop except handler), not UI. Moving it to interfaces/
would create a backward runtime→interface dependency. It stays in the runtime
cluster (→ reyn.runtime in PR-B). UI relocate = repl + renderer only.

Verification: broad cli/interfaces/repl/renderer sweep 1742 passed; the 1 fail
(test_eval_judge_..._artifact_handle_e2e) is a pre-existing env quirk
(reproduces on clean origin/main with changes stashed — falsified). ruff clean.
straggler grep (reyn.chat.repl / reyn.chat.renderer) = 0.

Refs #312
