+++ src/reyn/interfaces/repl/__init__.py
@@ -1,0 +1,5 @@
+"""Reyn REPL / console UI (the interactive chat front-end).
+
+#312 PR-A: relocated from ``reyn.chat`` (the agent runtime) — these are the UI
+modules #1700 left behind. ``reyn.chat`` itself becomes ``reyn.runtime`` in PR-B.
+"""
